Recover from exceptions querying GE export API (#35)
* Recover from exceptions querying GE export API

This is a migitation for https://github.com/gradle/bt-dev-prod-data-collector/issues/34

When the query fails, the default behavior was shutting down the app,
resulting in a lot of alerts. Now let's try to recover from such
exceptions until we found the root cause.

* Add retry logic

* lookupLimit:10000 -> lookupLimit:1000

We see a lot of timeout, let's try decrese this number.